Major rework of the merging strategy.
The merging logic has now been moved into the DISCOSNamespace class
itself. This has improved performance by a factor of 10, making
multiprocessing not needed anymore. The mean times from the reception of
a new message to the completion of the merging process into the main
dictionary has a value of approximately 0.3ms.